+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Correct filtering of sport events [#3986](https://github.com/rokwire/illinois-app/issues/3986).
 
 ## [5.1.32] - 2024-02-19
 ### Removed