+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-<!-- The next version number should be the version from pubspec.yaml - 5.2.0 -->
+## [5.2.0] - 2024-02-20
 ### Fixed
 - Correct filtering of sport events [#3986](https://github.com/rokwire/illinois-app/issues/3986).
 ### Changed